pyproject.toml
[project]
name = "theeni"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg-pool>=3.2.6",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
# "E" (pycodestyle errors), "F" (Pyflakes), "W" (pycodestyle warnings), "I" (isort)
select = ["E", "F", "W", "I"]

[tool.ruff.format]
# We're using ruff's own formatter
quote-style = "double"
indent-style = "space"
